refactor: Configuration moved to lua functions
This changes the way that nightfox is configured. When create the
example of overriding highlight groups there was an ordering of
configuration steps that was highlighted.

This change converts the global variables that used to configure
nightfox into a lua setup function. The loading of this configuration
and applying the colorscheme is in a seperate load function.

This change also documents this in the readme.

The old global variable configuration is checked for and deprication
warning are shown.

BREAKING CHANGE