fix(dayfox): Darken dayfox colors for readability
This change updates dayfox's colors to help with readability. The list
of changes to achive this are as follows:

- Darken the bg color and make the white color uniform in rgb space
- Darken colors to make them more visible. The colors that changed are:
  - red
  - green
  - yellow
  - cyan
  - orange

With regards to readability the two colors that are the worst culprits
are `yellow` and `orange`.

In the theme file there are some colors set to bright variants. Now
depending on the background color these will either be bright for dark
or dim for light.

Closes: #47