[Nvim-Tree] Configure a transparent background if 'transparent' is set (#66)
Currently, if the transparent option is set, Nvim Tree will still use
the solid background color configured in the colors table. As it stands
there is no way to make so that the side window matches the transparency
of the main window.

This change proposes a change targeted only to NvimTree. If accepted
this approach could be use more generally by replacing the c.bg_sidebar
with none and not having to check each time bg_sidebar is used.