Rewrite (#77)
The goal of this rewrite is to define a set of colors for each style (`pallets`). These styles then are mapped to a
specification (`spec`) that is used by the groups/modules. Users can override a color pallet or just a spec mapping. For
example you would like to just change functions to be `cyan` instead of blue. You can override just this and still have
everything else the same.

With all these configuration settings it is also important to have a way to precompute the results and cache the results
to make sure startup times are unaffected. This rewrite adds the concept of `compiled` files that are the vim calls
required to set the colorscheme. Nightfox also comes with its default settings precompiled to speed up startup time.

List of Breaking Changes:

- `nightfox.load()` has been removed in favour of setting colorschemes using the builtin `:colorscheme` command
- `nightfox.setup()` function now takes configuration options as an optional key
- `fox` has been removed in favour of setting colorschemes using the builtin `:colorscheme` command
- `transparent` has been moved to `options.transparent`
- `terminal_color` has been moved to `options.terminal_color`
- `alt_nc` has been moved to `options.dim_inactive`
- `styles` has been moved to `options.styles`
- `inverse` has been moved to `options.inverse`
- `colors` has been replaced with the concept of pallets
- `hlgroups` has been replaced with the concept of groups

----

* chore: Delete old files

* docs: Initial concept of api design

* feat(api): Construct a skeleton api and override store

* feat(config): Add base config and init funciton

* feat(lib): Add color lib object

* feat(lib): Add Shade lib

* fix(lib): Cleanup color lib

* feat(pallet): Add all pallets and spec definitions

* feat(group): Add base groups

* feat(color): Add color scripts for `:colorscheme` cmd

* refactor: Move main loading to main script

* feat(group): Override group with templates

* feat(group): Add plugin modules

* feat(lib): Add deprication warnings for config

* feat(statusline): Add lualine

* chore: Fix missing option rename

* refactor: Store hex string instead of Color objects

* feat(lib): Color shorthand construct from __call

* refactor: Deprecate `.load()` in favour of builtin

BREAKING CHANGE: Deprecate `.load()` in favour of builtin

* Add module option to toggle background

* feat(template): Change spec to use pallet as template base

* fix(lib): Highlight now does not link if set to ""

* feat(util): Add deprication warning for all old configs

BREAKING CHANGE: .setup() function has been changed. All settings have
been moved under `options`.

* fix(config): Rename alt_nc to dim_inactive

* refactor(template): Remove `${}` from template

* docs: New readme with detailed example usage

* chore: Fix spelling mistake for Deprecated

* docs(help): Start on help file

* fix: Fix wrong git spec

* feat(lib): Add compile lib

* feat(lib): Add precompiled functionality

* feat(lib): Add lua script to precompile base config

* feat(main): Use precompiled if no override or config

* feat(compile): Add `NightfoxCompile` command

* docs: Update readme for compile

* chore: Add .envrc file

* refactor(compile): Compile now sourced with `luafile`

Precompiled files are now sourced with the command `luafile`. This lets
the precompiled files live outside of the lua runtimepath. This means
that the default location of the precompiled files can be the system's
cache directory instead of neovim's config/lua folder.

Since the file is now executed by sourcing, there is no way to passing
which style to use. This meant that the precompiled file needed to be
broken up into one file for each style.

* docs(usage): Write usage/help doc

* docs: Add contributing docs

* chore: Remove scratch design file

* ci: Update doc action to use usage file

* ci: Pr check for stylua formatting

* chore: Change `colors/` to vim files

`vim` does not check for `.lua` file. This lets vim also set
colorschemes.

* feat(extra): Add extra generation

* feat(compile): Expand compiled file to one vim.cmd

This shaves off about 0.5 ms (2.086 -> 1.625) from the startup time.

* fix(statusline): lualine using Color lib correctly

* docs: Add type docs

* style(compile): Stylua compatible compile output

* style: color hex strings are lowercase

* feat(compile) Add status command to view compiled files

* refactor(main): Cleanup setting terminal function

* docs: Validate readme examples

* feat(module): Add symbol-outline

* fix(util): Deprication of missing config components

* docs: Add notice about breaking change

* fix(make): Change extra recipe to extragen

This make it so that the recipe is not the same as the extra folder.

* fix(extra): Correct url raw download

* style: fix tabs in stylua ignore block

* chore: Change readme images to v2 examples

* docs(usage): Duplicate to_hex tag

* docs: Link to rewrite pr

* docs: Update example screenshots

* docs: Fix incorrect links

* docs: Add migration table from v1 to v2 colors