Fix/vim compat (#109)
* chore(nix): Add vim derivation with lua support

* fix(vim): Guard `stdpath` as neovim only api

* fix(vim): Abstract calling vim.cmd

In neovim it is `vim.cmd` and in vim it is `vim.command` because of
course they are different....

* fix(vim): Abstract `bit` library for neovim and vim

`bit` is from `luajit` that is used by neovim. `bit32` is taken from lua
`5.2`. This also makes a requirement that vim's lua version is at least
`5.2`. This was added to requirements section of the docs.