feat(compile): Compile to lua byte-code
This change changes the compiler from compiling to lua script files to
the actual lua bytecode. This significantly speeds up nightfox.

This also changes the compiler from an addon to how nightfox operates.
This means that the compiler is now automatic and the user does not have
to handle it themselves.

This change was based off of the compiler changes made in catppuccin.
They implementend nightfox's compiler concept and improved its
performance. Thanks catppuccin for the great concept.