Feat/auto compile (#241)
* feat(compile): Compile to lua byte-code

This change changes the compiler from compiling to lua script files to
the actual lua bytecode. This significantly speeds up nightfox.

This also changes the compiler from an addon to how nightfox operates.
This means that the compiler is now automatic and the user does not have
to handle it themselves.

This change was based off of the compiler changes made in catppuccin.
They implementend nightfox's compiler concept and improved its
performance. Thanks catppuccin for the great concept.

* feat(log): Add logging for vim and neovim

* fix(vim): Support lua 5.3

Lua lib `bit` and `bit32` is not present in 5.3. These were added to the
native language. This commit adds a native version of bit for 5.3.

* chore: ignore native binopts file

* fix(lib): Fix missing renames

* fix(vim): loadstring is load in lua 5.2 and above

* fix(lib): Convert rgb float values to int b4 left shift

* fix: Remove redundant vim folder for cache dir

* perf(compile): Remove protected call to compiled func