chore: remove precommit hooks
Fingerprint generation was getting in the way of every branch as there
would be merge conflicts with every change. This way how it will be part
of the pr workflow and when a new change in main is introduced then
there will be an automatic new commit with a new fingerprint.