Fix failing compat code around a deprecation.
Both load and loadstring exist in both liblua 5.1 and liblua 5.2,
but load doesn't accept strings in 5.1. So this wasn't backwards
compatible.

For now, try the older one first, assume it'll be removed in a
future version, and hope it doesn't break Vims compiled with liblua
5.2 that I've not tested.

Fixes an error: "bad argument #1 to 'ld' (function expected, got string)"
in liblua 5.1.